+++ libuproc/alphabet.c
@@ -54,8 +54,8 @@ uproc_alphabet_create(const char *s)
 
     if (strlen(s) != UPROC_ALPHABET_SIZE) {
         uproc_error_msg(
-            UPROC_EINVAL, "string too short: %zu characters instead of %d",
-            strlen(s), UPROC_ALPHABET_SIZE);
+            UPROC_EINVAL, "string too short: %lu characters instead of %d",
+            (unsigned long)strlen(s), UPROC_ALPHABET_SIZE);
         goto error;
     }
     strcpy(a->str, s);
